Await initialization before connecting
